update: ファイルに関係したコマンドを別のファイルにまとめた
